# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Tekton Hub
    description: HTTP services for managing Tekton Hub
    version: "1.0"
servers:
    - url: https://api.hub.tekton.dev
paths:
...
    /system/config/refresh:
        post:
            tags:
                - admin
            summary: RefreshConfig admin
            description: Refresh the changes in config file
            operationId: admin#RefreshConfig
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RefreshConfigResponseBody'
                            example:
                                checksum: 41ba391c8baf1fcd3c62c11272b913dc6613f4cf3b1833cfbb32431dc4384c93
                "401":
                    description: 'invalid-token: Invalid User token'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "403":
                    description: 'invalid-scopes: Invalid Token scopes '
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal-error: Internal server error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
            security:
                - jwt_header_Authorization:
                    - rating:read
                    - rating:write
                    - agent:create
                    - catalog:refresh
                    - config:refresh
                    - refresh:token
    /system/user/agent:
        put:
            tags:
                - admin
            summary: UpdateAgent admin
            description: Create or Update an agent user with required scopes
            operationId: admin#UpdateAgent
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAgentRequestBody'
                        example:
                            name: abc
                            scopes:
                                - catalog-refresh
                                - agent:create
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateAgentResponseBody'
                            example:
                                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1Nzc4ODAzMDAsImlhdCI6MTU3Nzg4MDAwMCwiaWQiOjExLCJpc3MiOiJUZWt0b24gSHViIiwic2NvcGVzIjpbInJhdGluZzpyZWFkIiwicmF0aW5nOndyaXRlIiwiYWdlbnQ6Y3JlYXRlIl0sInR5cGUiOiJhY2Nlc3MtdG9rZW4ifQ.6pDmziSKkoSqI1f0rc4-AqVdcfY0Q8wA-tSLzdTCLgM
                "400":
                    description: 'invalid-payload: Invalid request body'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "401":
                    description: 'invalid-token: Invalid User token'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "403":
                    description: 'invalid-scopes: Invalid Token scopes '
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal-error: Internal server error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
            security:
                - jwt_header_Authorization:
                    - rating:read
                    - rating:write
                    - agent:create
                    - catalog:refresh
                    - config:refresh
                    - refresh:token
...
components:
    schemas:
...
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: false
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: false
            description: Invalid request body
            example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
...
        RefreshConfigResponseBody:
            type: object
            properties:
                checksum:
                    type: string
                    description: Config file checksum
                    example: 41ba391c8baf1fcd3c62c11272b913dc6613f4cf3b1833cfbb32431dc4384c93
            example:
                checksum: 41ba391c8baf1fcd3c62c11272b913dc6613f4cf3b1833cfbb32431dc4384c93
            required:
                - checksum
...
        UpdateAgentRequestBody:
            type: object
            properties:
                name:
                    type: string
                    description: Name of Agent
                    example: abc
                scopes:
                    type: array
                    items:
                        type: string
                        example: Quia eos ipsum aperiam nam numquam rerum.
                    description: Scopes required for Agent
                    example:
                        - catalog-refresh
                        - agent:create
            example:
                name: abc
                scopes:
                    - catalog-refresh
                    - agent:create
            required:
                - name
                - scopes
        UpdateAgentResponseBody:
            type: object
            properties:
                token:
                    type: string
                    description: Agent JWT
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1Nzc4ODAzMDAsImlhdCI6MTU3Nzg4MDAwMCwiaWQiOjExLCJpc3MiOiJUZWt0b24gSHViIiwic2NvcGVzIjpbInJhdGluZzpyZWFkIiwicmF0aW5nOndyaXRlIiwiYWdlbnQ6Y3JlYXRlIl0sInR5cGUiOiJhY2Nlc3MtdG9rZW4ifQ.6pDmziSKkoSqI1f0rc4-AqVdcfY0Q8wA-tSLzdTCLgM
            example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1Nzc4ODAzMDAsImlhdCI6MTU3Nzg4MDAwMCwiaWQiOjExLCJpc3MiOiJUZWt0b24gSHViIiwic2NvcGVzIjpbInJhdGluZzpyZWFkIiwicmF0aW5nOndyaXRlIiwiYWdlbnQ6Y3JlYXRlIl0sInR5cGUiOiJhY2Nlc3MtdG9rZW4ifQ.6pDmziSKkoSqI1f0rc4-AqVdcfY0Q8wA-tSLzdTCLgM
            required:
                - token
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Secures endpoint by requiring a valid JWT retrieved via the /auth/login endpoint.
            scheme: bearer
...